rename entropy implementation details to thermodynamics
