+++ src/hooks/index.ts
@@ -1,4 +1,4 @@
+export { useThermodynamics } from "./use-thermodynamics"
 export { useEntropy } from "./use-entropy"
-export { useEntropyContext } from "./use-entropy-context"
 export { useInitialConditions } from "./use-initial-conditions"
 export { usePhysics } from "./use-physics"
